Update social graph service test DBManager import
