+++ examples/social_graph_bot.py
@@ -10,18 +10,21 @@ import uuid
 from datetime import timedelta, timezone
 from typing import TYPE_CHECKING, List, Tuple
 
+import aiohttp
+
 from deepthought.goal_scheduler import GoalScheduler
 from deepthought.graph.connector import GraphConnector
 from deepthought.graph.dal import GraphDAL
-from deepthought.services.file_graph_dal import FileGraphDAL
-
-
-
-import aiohttp
-import aiosqlite
-
 from deepthought.services import PersonaManager
-
+from deepthought.services.db_manager import (
+    MAX_MEMORY_LENGTH,
+    MAX_PROMPT_LENGTH,
+    MAX_THEORY_LENGTH,
+    DBManager,
+)
+from deepthought.services.file_graph_dal import FileGraphDAL
+from deepthought.services.moderation import is_allowed
+from deepthought.services.scheduler import SchedulerService
 
 try:
     import discord
@@ -209,564 +212,6 @@ async def generate_idle_response(prompt: str | None = None) -> str | None:
 
 # Simple list of phrases considered bullying
 BULLYING_PHRASES = ["idiot", "stupid", "loser", "dumb", "ugly"]
-
-# Limits used when validating inputs
-MAX_MEMORY_LENGTH = 1000
-MAX_THEORY_LENGTH = 256
-MAX_PROMPT_LENGTH = 2000
-
-class DBManager:
-    """Lightweight wrapper managing a single aiosqlite connection."""
-
-    CREATE_TABLE_QUERIES = [
-        """
-        CREATE TABLE IF NOT EXISTS interactions (
-            user_id TEXT,
-            target_id TEXT,
-            timestamp DATETIME DEFAULT CURRENT_TIMESTAMP
-        )
-        """,
-        """
-        CREATE TABLE IF NOT EXISTS affinity (
-            user_id TEXT PRIMARY KEY,
-            score INTEGER DEFAULT 0
-        )
-        """,
-        """
-        CREATE TABLE IF NOT EXISTS relationships (
-            source_id TEXT,
-            target_id TEXT,
-            interaction_count INTEGER DEFAULT 0,
-            sentiment_sum REAL DEFAULT 0,
-            PRIMARY KEY(source_id, target_id)
-        )
-        """,
-        """
-        CREATE TABLE IF NOT EXISTS memories (
-            user_id TEXT,
-            topic TEXT,
-            memory TEXT,
-            sentiment_score REAL,
-            timestamp DATETIME DEFAULT CURRENT_TIMESTAMP
-        )
-        """,
-        """
-        CREATE TABLE IF NOT EXISTS theories (
-            subject_id TEXT,
-            theory TEXT,
-            confidence REAL,
-            updated TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-            PRIMARY KEY(subject_id, theory)
-        )
-        """,
-        """
-        CREATE TABLE IF NOT EXISTS queued_tasks (
-            task_id INTEGER PRIMARY KEY,
-            user_id TEXT,
-            context TEXT,
-            prompt TEXT,
-            status TEXT DEFAULT 'pending',
-            created TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-        )
-        """,
-        """
-        CREATE TABLE IF NOT EXISTS summary_goals (
-            task_id INTEGER PRIMARY KEY,
-            user_id TEXT,
-            context TEXT,
-            prompt TEXT,
-            status TEXT DEFAULT 'pending',
-            created TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-        )
-        """,
-        """
-        CREATE TABLE IF NOT EXISTS sentiment_trends (
-            user_id TEXT,
-            channel_id TEXT,
-            sentiment_sum REAL DEFAULT 0,
-            message_count INTEGER DEFAULT 0,
-            PRIMARY KEY(user_id, channel_id)
-        )
-        """,
-        """
-        CREATE TABLE IF NOT EXISTS themes (
-            user_id TEXT,
-            channel_id TEXT,
-            theme TEXT,
-            updated TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-            PRIMARY KEY(user_id, channel_id)
-        )
-        """,
-        """
-        CREATE TABLE IF NOT EXISTS user_flags (
-            user_id TEXT PRIMARY KEY,
-            do_not_mock INTEGER
-        )
-        """,
-        """
-        CREATE TABLE IF NOT EXISTS recent_topics (
-            topic TEXT PRIMARY KEY,
-            last_used TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-        )
-        """,
-    ]
-
-    def __init__(self, db_path: str = DB_PATH) -> None:
-        self.db_path = db_path
-        self._db: aiosqlite.Connection | None = None
-
-    async def connect(self) -> None:
-        if self._db is None:
-            dir_path = os.path.dirname(self.db_path)
-            if dir_path:
-                os.makedirs(dir_path, exist_ok=True)
-            self._db = await aiosqlite.connect(self.db_path)
-
-    async def close(self) -> None:
-        if self._db is not None:
-            await self._db.close()
-            self._db = None
-
-    def _create_table_statements(self) -> list[str]:
-        """Return SQL statements for creating required tables."""
-        try:
-            return CREATE_TABLE_QUERIES
-        except NameError:  # pragma: no cover - executed in isolated test context
-            return [
-                """
-                CREATE TABLE IF NOT EXISTS interactions (
-                    user_id TEXT,
-                    target_id TEXT,
-                    timestamp DATETIME DEFAULT CURRENT_TIMESTAMP
-                )
-                """,
-                """
-                CREATE TABLE IF NOT EXISTS affinity (
-                    user_id TEXT PRIMARY KEY,
-                    score INTEGER DEFAULT 0
-                )
-                """,
-                """
-                CREATE TABLE IF NOT EXISTS relationships (
-                    source_id TEXT,
-                    target_id TEXT,
-                    interaction_count INTEGER DEFAULT 0,
-                    sentiment_sum REAL DEFAULT 0,
-                    PRIMARY KEY(source_id, target_id)
-                )
-                """,
-                """
-                CREATE TABLE IF NOT EXISTS memories (
-                    user_id TEXT,
-                    topic TEXT,
-                    memory TEXT,
-                    sentiment_score REAL,
-                    timestamp DATETIME DEFAULT CURRENT_TIMESTAMP
-                )
-                """,
-                """
-                CREATE TABLE IF NOT EXISTS theories (
-                    subject_id TEXT,
-                    theory TEXT,
-                    confidence REAL,
-                    updated TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                    PRIMARY KEY(subject_id, theory)
-                )
-                """,
-                """
-                CREATE TABLE IF NOT EXISTS queued_tasks (
-                    task_id INTEGER PRIMARY KEY,
-                    user_id TEXT,
-                    context TEXT,
-                    prompt TEXT,
-                    status TEXT DEFAULT 'pending',
-                    created TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-                )
-                """,
-                """
-                CREATE TABLE IF NOT EXISTS sentiment_trends (
-                    user_id TEXT,
-                    channel_id TEXT,
-                    sentiment_sum REAL DEFAULT 0,
-                    message_count INTEGER DEFAULT 0,
-                    PRIMARY KEY(user_id, channel_id)
-                )
-                """,
-                """
-                CREATE TABLE IF NOT EXISTS themes (
-                    user_id TEXT,
-                    channel_id TEXT,
-                    theme TEXT,
-                    updated TIMESTAMP DEFAULT CURRENT_TIMESTAMP,
-                    PRIMARY KEY(user_id, channel_id)
-                )
-                """,
-                """
-                CREATE TABLE IF NOT EXISTS user_flags (
-                    user_id TEXT PRIMARY KEY,
-                    do_not_mock INTEGER
-                )
-                """,
-                """
-                CREATE TABLE IF NOT EXISTS recent_topics (
-                    topic TEXT PRIMARY KEY,
-                    last_used TIMESTAMP DEFAULT CURRENT_TIMESTAMP
-                )
-                """,
-            ]
-
-    async def init_db(self) -> None:
-        await self.connect()
-        assert self._db
-        for query in self.CREATE_TABLE_QUERIES:
-
-            await self._db.execute(query)
-
-        await self._db.commit()
-
-    async def log_interaction(
-        self,
-        user_id: int,
-        target_id: int | None = None,
-        sentiment_score: float | None = None,
-    ) -> None:
-        await self.connect()
-        assert self._db
-        if sentiment_score is not None:
-            if not isinstance(sentiment_score, (int, float)):
-                raise ValueError("sentiment_score must be numeric")
-            if not -1 <= float(sentiment_score) <= 1:
-                raise ValueError("sentiment_score out of range")
-        await self._db.execute(
-            "INSERT INTO interactions (user_id, target_id) VALUES (?, ?)",
-            (str(user_id), str(target_id) if target_id is not None else None),
-        )
-        await self._db.execute(
-            """
-            INSERT INTO affinity (user_id, score)
-            VALUES (?, 1)
-            ON CONFLICT(user_id) DO UPDATE SET score=affinity.score + 1
-            """,
-            (str(user_id),),
-        )
-        if target_id is not None:
-            await self._db.execute(
-                """
-                INSERT INTO relationships (source_id, target_id, interaction_count, sentiment_sum)
-                VALUES (?, ?, 1, ?)
-                ON CONFLICT(source_id, target_id) DO UPDATE SET
-                    interaction_count=relationships.interaction_count + 1,
-                    sentiment_sum=relationships.sentiment_sum + excluded.sentiment_sum
-                """,
-                (str(user_id), str(target_id), sentiment_score or 0.0),
-            )
-        await self._db.commit()
-
-    async def recall_user(self, user_id: int):
-        await self.connect()
-        assert self._db
-        async with self._db.execute(
-            "SELECT topic, memory FROM memories WHERE user_id= ?",
-            (str(user_id),),
-        ) as cur:
-            return await cur.fetchall()
-
-    async def store_memory(
-        self,
-        user_id: int,
-        memory: str,
-        topic: str = "",
-        sentiment_score: float | None = None,
-    ) -> None:
-        if not isinstance(memory, str) or not memory.strip():
-            raise ValueError("memory must be a non-empty string")
-        if len(memory) > MAX_MEMORY_LENGTH:
-            raise ValueError("memory exceeds maximum length")
-        if not isinstance(topic, str):
-            raise ValueError("topic must be a string")
-        if sentiment_score is not None:
-            if not isinstance(sentiment_score, (int, float)):
-                raise ValueError("sentiment_score must be numeric")
-            if not -1 <= float(sentiment_score) <= 1:
-                raise ValueError("sentiment_score out of range")
-
-        await self.connect()
-        assert self._db
-        await self._db.execute(
-            "INSERT INTO memories (user_id, topic, memory, sentiment_score) VALUES (?, ?, ?, ?)",
-            (str(user_id), topic, memory, sentiment_score),
-        )
-        if topic:
-            await self._db.execute(
-                """
-                INSERT INTO recent_topics (topic, last_used)
-                VALUES (?, CURRENT_TIMESTAMP)
-                ON CONFLICT(topic) DO UPDATE SET last_used=CURRENT_TIMESTAMP
-                """,
-                (topic,),
-            )
-        await self._db.commit()
-
-    async def store_theory(self, subject_id: int, theory: str, confidence: float) -> None:
-        if not isinstance(theory, str) or not theory.strip():
-            raise ValueError("theory must be a non-empty string")
-        if len(theory) > MAX_THEORY_LENGTH:
-            raise ValueError("theory exceeds maximum length")
-        if not isinstance(confidence, (int, float)):
-            raise ValueError("confidence must be numeric")
-        if not 0 <= float(confidence) <= 1:
-            raise ValueError("confidence out of range")
-
-        await self.connect()
-        assert self._db
-        await self._db.execute(
-            """
-            INSERT INTO theories (subject_id, theory, confidence)
-            VALUES (?, ?, ?)
-            ON CONFLICT(subject_id, theory) DO UPDATE SET
-                confidence=excluded.confidence,
-                updated=CURRENT_TIMESTAMP
-            """,
-            (str(subject_id), theory, confidence),
-        )
-        await self._db.commit()
-
-    async def get_theories(self, subject_id: int):
-        await self.connect()
-        assert self._db
-        async with self._db.execute(
-            "SELECT theory, confidence FROM theories WHERE subject_id=?",
-            (str(subject_id),),
-        ) as cur:
-            return await cur.fetchall()
-
-    async def update_sentiment_trend(
-        self,
-        user_id: int,
-        channel_id: int,
-        sentiment_score: float,
-    ) -> None:
-        if not isinstance(sentiment_score, (int, float)):
-            raise ValueError("sentiment_score must be numeric")
-        if not -1 <= float(sentiment_score) <= 1:
-            raise ValueError("sentiment_score out of range")
-        await self.connect()
-        assert self._db
-        await self._db.execute(
-            """
-            INSERT INTO sentiment_trends (user_id, channel_id, sentiment_sum, message_count)
-            VALUES (?, ?, ?, 1)
-            ON CONFLICT(user_id, channel_id) DO UPDATE SET
-                sentiment_sum=sentiment_trends.sentiment_sum + excluded.sentiment_sum,
-                message_count=sentiment_trends.message_count + 1
-            """,
-            (str(user_id), str(channel_id), sentiment_score),
-        )
-        await self._db.commit()
-
-    async def get_sentiment_trend(self, user_id: int, channel_id: int):
-        await self.connect()
-        assert self._db
-        async with self._db.execute(
-            "SELECT sentiment_sum, message_count FROM sentiment_trends WHERE user_id=? AND channel_id=?",
-            (str(user_id), str(channel_id)),
-        ) as cur:
-            return await cur.fetchone()
-
-    async def queue_deep_reflection(self, user_id: int, context: dict, prompt: str) -> int:
-
-        if not isinstance(prompt, str) or not prompt.strip():
-            raise ValueError("prompt must be a non-empty string")
-        if len(prompt) > MAX_PROMPT_LENGTH:
-            raise ValueError("prompt exceeds maximum length")
-        if not isinstance(context, dict):
-            raise ValueError("context must be a dictionary")
-        try:
-            context_json = json.dumps(context)
-        except (TypeError, ValueError) as exc:
-            raise ValueError("context is not JSON serializable") from exc
-
-        await self.connect()
-        assert self._db
-        cur = await self._db.execute(
-            "INSERT INTO queued_tasks (user_id, context, prompt) VALUES (?, ?, ?)",
-            (str(user_id), context_json, prompt),
-        )
-        await self._db.commit()
-        return cur.lastrowid
-
-    async def list_pending_tasks(self):
-        """Return pending reflection tasks."""
-        await self.connect()
-        assert self._db
-        async with self._db.execute(
-            "SELECT task_id, user_id, context, prompt FROM queued_tasks WHERE status='pending'"
-        ) as cur:
-            return await cur.fetchall()
-
-    async def mark_task_done(self, task_id: int) -> None:
-        """Mark a queued task as completed."""
-        await self.connect()
-        assert self._db
-        await self._db.execute(
-            "UPDATE queued_tasks SET status='done' WHERE task_id=?",
-            (task_id,),
-        )
-        await self._db.commit()
-
-    async def add_summary_goal(self, user_id: int, context: dict, prompt: str) -> int:
-        """Store a generated summary and goal using the queued task schema."""
-        if not isinstance(prompt, str) or not prompt.strip():
-            raise ValueError("prompt must be a non-empty string")
-        if not isinstance(context, dict):
-            raise ValueError("context must be a dictionary")
-        try:
-            context_json = json.dumps(context)
-        except (TypeError, ValueError) as exc:
-            raise ValueError("context is not JSON serializable") from exc
-
-        await self.connect()
-        assert self._db
-        cur = await self._db.execute(
-            "INSERT INTO summary_goals (user_id, context, prompt) VALUES (?, ?, ?)",
-            (str(user_id), context_json, prompt),
-        )
-        await self._db.commit()
-        return cur.lastrowid
-
-    async def list_pending_summary_goals(self):
-        """Return pending summary/goal tasks."""
-        await self.connect()
-        assert self._db
-        async with self._db.execute(
-            "SELECT task_id, user_id, context, prompt FROM summary_goals WHERE status='pending'"
-        ) as cur:
-            return await cur.fetchall()
-
-    async def mark_summary_goal_done(self, task_id: int) -> None:
-        """Mark a stored summary goal as completed."""
-        await self.connect()
-        assert self._db
-        await self._db.execute(
-            "UPDATE summary_goals SET status='done' WHERE task_id=?",
-            (task_id,),
-        )
-        await self._db.commit()
-
-    async def set_do_not_mock(self, user_id: int, flag: bool = True) -> None:
-        await self.connect()
-        assert self._db
-        await self._db.execute(
-            """
-            INSERT INTO user_flags (user_id, do_not_mock)
-            VALUES (?, ?)
-            ON CONFLICT(user_id) DO UPDATE SET do_not_mock=excluded.do_not_mock
-            """,
-            (str(user_id), int(flag)),
-        )
-        await self._db.commit()
-
-    async def is_do_not_mock(self, user_id: int) -> bool:
-        await self.connect()
-        assert self._db
-        async with self._db.execute(
-            "SELECT do_not_mock FROM user_flags WHERE user_id=?",
-            (str(user_id),),
-        ) as cur:
-            row = await cur.fetchone()
-            return bool(row[0]) if row else False
-
-    async def adjust_affinity(self, user_id: int, delta: int) -> None:
-        await self.connect()
-        assert self._db
-        await self._db.execute(
-            """
-            INSERT INTO affinity (user_id, score)
-            VALUES (?, ?)
-            ON CONFLICT(user_id) DO UPDATE SET score=affinity.score + ?
-            """,
-            (str(user_id), delta, delta),
-        )
-        await self._db.commit()
-
-    async def get_affinity(self, user_id: int) -> int:
-        await self.connect()
-        assert self._db
-        async with self._db.execute(
-            "SELECT score FROM affinity WHERE user_id=?",
-            (str(user_id),),
-        ) as cur:
-            row = await cur.fetchone()
-            return int(row[0]) if row else 0
-
-    async def get_relationship(self, user_id: int, target_id: int):
-        await self.connect()
-        assert self._db
-        async with self._db.execute(
-            "SELECT interaction_count, sentiment_sum FROM relationships WHERE source_id=? AND target_id=?",
-            (str(user_id), str(target_id)),
-        ) as cur:
-            return await cur.fetchone()
-
-    async def _get_relationship_avg(self, user_id: int, target_id: int) -> float:
-        row = await self.get_relationship(user_id, target_id)
-        if not row or not row[0]:
-            return 0.0
-        count, sentiment_sum = row
-        return float(sentiment_sum) / count
-
-    async def get_friendliness(self, user_id: int, target_id: int) -> float:
-        avg = await self._get_relationship_avg(user_id, target_id)
-        return max(0.0, avg)
-
-    async def get_hostility(self, user_id: int, target_id: int) -> float:
-        avg = await self._get_relationship_avg(user_id, target_id)
-        return min(0.0, avg)
-
-    async def set_theme(self, user_id: int, channel_id: int, theme: str) -> None:
-        if not isinstance(theme, str) or not theme.strip():
-            raise ValueError("theme must be a non-empty string")
-        await self.connect()
-        assert self._db
-        await self._db.execute(
-            """
-            INSERT INTO themes (user_id, channel_id, theme)
-            VALUES (?, ?, ?)
-            ON CONFLICT(user_id, channel_id) DO UPDATE SET
-                theme=excluded.theme,
-                updated=CURRENT_TIMESTAMP
-            """,
-            (str(user_id), str(channel_id), theme),
-        )
-        await self._db.commit()
-
-    async def get_theme(self, user_id: int, channel_id: int):
-        await self.connect()
-        assert self._db
-        async with self._db.execute(
-            "SELECT theme FROM themes WHERE user_id=? AND channel_id=?",
-            (str(user_id), str(channel_id)),
-        ) as cur:
-            row = await cur.fetchone()
-            return row[0] if row else None
-
-    async def get_all_sentiment_trends(self):
-        await self.connect()
-        assert self._db
-        async with self._db.execute(
-            "SELECT user_id, channel_id, sentiment_sum, message_count FROM sentiment_trends"
-        ) as cur:
-            return await cur.fetchall()
-
-    async def get_recent_topics(self, limit: int = 3) -> list[str]:
-        await self.connect()
-        assert self._db
-        async with self._db.execute(
-            "SELECT topic FROM recent_topics ORDER BY last_used DESC LIMIT ?",
-            (limit,),
-        ) as cur:
-            rows = await cur.fetchall()
-            return [r[0] for r in rows]
 
 
 DEFAULT_DB_PATH = DB_PATH
