+++ src/deepthought/services/__init__.py
@@ -1,5 +1,6 @@
 """Service utilities for DeepThought."""
 
+from .db_manager import DBManager
 from .file_graph_dal import FileGraphDAL
 from .hierarchical_service import HierarchicalService
 from .memory_service import MemoryService
@@ -12,5 +13,5 @@ __all__ = [
     "HierarchicalService",
     "PersonaManager",
     "SocialGraphService",
-
+    "DBManager",
 ]
