+++ tests/test_bot_cleanup.py
@@ -4,6 +4,7 @@ import discord
 import pytest
 
 import examples.social_graph_bot as sg
+from deepthought.services import DBManager
 
 
 class DummyNATS:
@@ -18,7 +19,7 @@ class DummyNATS:
 
 @pytest.mark.asyncio
 async def test_bot_cleanup_on_cancel(tmp_path, monkeypatch):
-    sg.db_manager = sg.DBManager(str(tmp_path / "sg.db"))
+    sg.db_manager = DBManager(str(tmp_path / "sg.db"))
     await sg.db_manager.connect()
     await sg.db_manager.init_db()
