+++ tests/test_db_manager_init.py
@@ -1,9 +1,7 @@
-import ast
-import types
-from pathlib import Path
-
 import aiosqlite
 import pytest
+
+from deepthought.services import DBManager
 
 TABLES = [
     "interactions",
@@ -18,25 +16,8 @@ TABLES = [
 ]
 
 
-def load_dbmanager():
-    path = Path(__file__).resolve().parents[1] / "examples" / "social_graph_bot.py"
-    source = path.read_text()
-    tree = ast.parse(source)
-    namespace = {
-        "aiosqlite": aiosqlite,
-        "os": __import__("os"),
-        "DB_PATH": str(path),
-    }
-    for node in tree.body:
-        if isinstance(node, ast.ClassDef) and node.name == "DBManager":
-            exec(compile(ast.Module([node], []), filename=str(path), mode="exec"), namespace)
-            return namespace["DBManager"]
-    raise RuntimeError("DBManager not found")
-
-
 @pytest.mark.asyncio
 async def test_db_manager_init_creates_tables_once(tmp_path):
-    DBManager = load_dbmanager()
     db_file = tmp_path / "db.sqlite"
     manager = DBManager(str(db_file))
     await manager.init_db()
