+++ tests/test_persona_integration.py
@@ -4,7 +4,7 @@ import random
 import pytest
 
 import examples.social_graph_bot as sg
-from deepthought.services import PersonaManager
+from deepthought.services import DBManager, PersonaManager
 
 
 class DummyAuthor:
@@ -54,7 +54,7 @@ class DummyMessage:
 
 @pytest.mark.asyncio
 async def test_on_message_persona_changes_with_affinity(tmp_path, monkeypatch, input_events):
-    sg.db_manager = sg.DBManager(str(tmp_path / "sg.db"))
+    sg.db_manager = DBManager(str(tmp_path / "sg.db"))
     await sg.db_manager.connect()
     await sg.db_manager.init_db()
