+++ tests/test_user_flags.py
@@ -2,11 +2,12 @@ import aiosqlite
 import pytest
 
 import examples.social_graph_bot as sg
+from deepthought.services import DBManager
 
 
 @pytest.mark.asyncio
 async def test_user_flags_table_and_functions(tmp_path):
-    sg.db_manager = sg.DBManager(str(tmp_path / "sg.db"))
+    sg.db_manager = DBManager(str(tmp_path / "sg.db"))
     await sg.db_manager.connect()
     await sg.db_manager.init_db()
